+++ src/main/java/com/plivo/api/models/message/MessageLister.java
@@ -3,8 +3,15 @@ package com.plivo.api.models.message;
 import com.plivo.api.models.base.ListResponse;
 import com.plivo.api.models.base.Lister;
 import com.plivo.api.util.PropertyFilter;
+
+import com.plivo.api.exceptions.PlivoRestException;
+import java.io.IOException;
+
+import java.lang.reflect.Field;
+
 import java.util.Date;
 import retrofit2.Call;
+import retrofit2.Response;
 
 /**
  * Pending list request that can be executed by calling list();
@@ -63,6 +70,29 @@ public class MessageLister extends Lister<Message> {
     return this;
   }
 
+  @Override  
+  /**
+  * Actually list instances of the resource.
+  */
+  public ListResponse<Message> list() throws IOException, PlivoRestException {
+    validate();
+    Response<ListResponse<Message>> response = obtainCall().execute();
+
+    handleResponse(response);
+    
+    try 
+    {
+      Field meta = response.body().getClass().getDeclaredField("meta");
+      meta.setAccessible(true);
+      meta.set(response.body(), new MessageMeta(response.body().getMeta()));
+    }
+    catch(Exception e) 
+    {
+      e.printStackTrace();
+    }
+    
+    return response.body();
+  }
 
   @Override
   protected Call<ListResponse<Message>> obtainCall() {