+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [5.0.0](https://github.com/plivo/plivo-java/tree/v5.0.0) (2021-06-21)
+- getTotalCount() method removed for list MDR.
+- Moved get() method to PowerPackLister
+
+## [4.15.2](https://github.com/plivo/plivo-java/tree/v4.15.2) (2021-06-18)
+- Removed total_count field from meta data for list MDR response. getTotalCount() for list MDR will return 0.
+
 ## [4.15.1](https://github.com/plivo/plivo-java/tree/v4.15.1) (2021-04-22)
 - Bug Fix for adaptive powerpack feature.